# SOFTARCH11/ClinicDesctop.Data/OpenAPIs/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "ClinicService",
      "version": "1.0"
    },
    "paths": {
      "/api/Client/create": {
        "post": {
          "tags": [
            "Client"
          ],
          "operationId": "ClientCreate",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateClientRequest"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateClientRequest"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateClientRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Client/update": {
        "put": {
          "tags": [
            "Client"
          ],
          "operationId": "ClientUpdate",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateClientRequest"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateClientRequest"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateClientRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Client/delete": {
        "delete": {
          "tags": [
            "Client"
          ],
          "operationId": "ClientDelete",
          "parameters": [
            {
              "name": "clientId",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Client/get-all": {
        "get": {
          "tags": [
            "Client"
          ],
          "operationId": "ClientGetAll",
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Client"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Client"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Client"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/Client/get/{clientId}": {
        "get": {
          "tags": [
            "Client"
          ],
          "operationId": "ClientGetById",
          "parameters": [
            {
              "name": "clientId",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/Client"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Client"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Client"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Pet/create": {
        "post": {
          "tags": [
            "Pet"
          ],
          "operationId": "PetCreate",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePetRequest"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePetRequest"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePetRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Pet/update": {
        "put": {
          "tags": [
            "Pet"
          ],
          "operationId": "PetUpdate",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePetRequest"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePetRequest"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePetRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Pet/delete": {
        "delete": {
          "tags": [
            "Pet"
          ],
          "operationId": "ClientDelete",
          "parameters": [
            {
              "name": "petId",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Pet/get-all": {
        "get": {
          "tags": [
            "Pet"
          ],
          "operationId": "PetGetAll",
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Pet"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Pet"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Pet"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/Pet/get/{clientId}": {
        "get": {
          "tags": [
            "Pet"
          ],
          "operationId": "PetGetById",
          "parameters": [
            {
              "name": "petId",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "clientId",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/Pet"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Pet"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Pet"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Client": {
          "type": "object",
          "properties": {
            "clientId": {
              "type": "integer",
              "format": "int32"
            },
            "document": {
              "type": "string",
              "nullable": true
            },
            "surName": {
              "type": "string",
              "nullable": true
            },
            "firstName": {
              "type": "string",
              "nullable": true
            },
            "patronymic": {
              "type": "string",
              "nullable": true
            },
            "birthday": {
              "type": "string",
              "format": "date-time"
            }
          },
          "additionalProperties": false
        },
        "CreateClientRequest": {
          "type": "object",
          "properties": {
            "document": {
              "type": "string",
              "nullable": true
            },
            "surName": {
              "type": "string",
              "nullable": true
            },
            "firstName": {
              "type": "string",
              "nullable": true
            },
            "patronymic": {
              "type": "string",
              "nullable": true
            },
            "birthday": {
              "type": "string",
              "format": "date-time"
            }
          },
          "additionalProperties": false
        },
        "CreatePetRequest": {
          "type": "object",
          "properties": {
            "clientId": {
              "type": "integer",
              "format": "int32"
            },
            "name": {
              "type": "string",
              "nullable": true
            },
            "birthday": {
              "type": "string",
              "format": "date-time"
            }
          },
          "additionalProperties": false
        },
        "Pet": {
          "type": "object",
          "properties": {
            "petId": {
              "type": "integer",
              "format": "int32"
            },
            "clientId": {
              "type": "integer",
              "format": "int32"
            },
            "name": {
              "type": "string",
              "nullable": true
            },
            "birthday": {
              "type": "string",
              "format": "date-time"
            }
          },
          "additionalProperties": false
        },
        "UpdateClientRequest": {
          "type": "object",
          "properties": {
            "clientId": {
              "type": "integer",
              "format": "int32"
            },
            "document": {
              "type": "string",
              "nullable": true
            },
            "surName": {
              "type": "string",
              "nullable": true
            },
            "firstName": {
              "type": "string",
              "nullable": true
            },
            "patronymic": {
              "type": "string",
              "nullable": true
            },
            "birthday": {
              "type": "string",
              "format": "date-time"
            }
          },
          "additionalProperties": false
        },
        "UpdatePetRequest": {
          "type": "object",
          "properties": {
            "petId": {
              "type": "integer",
              "format": "int32"
            },
            "clientId": {
              "type": "integer",
              "format": "int32"
            },
            "name": {
              "type": "string",
              "nullable": true
            },
            "birthday": {
              "type": "string",
              "format": "date-time"
            }
          },
          "additionalProperties": false
        }
      }
    }
  }